# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FileApi",
    "description": "File Api V1",
    "license": {
      "name": "",
      "identifier": "MIT"
    },
    "version": "0.8"
  },
  "paths": {
    "/api/v1/template": {
      "post": {
        "tags": [
          "template::routes"
        ],
        "operationId": "upsert",
        "parameters": [
          {
            "name": "_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "title",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "templateType",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TemplateType"
            }
          },
          {
            "name": "templateContext",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Context"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "contentMediaType": "application/octet-stream"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Upsert a template",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateV2"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/template/find-by-context": {
      "get": {
        "tags": [
          "template::routes"
        ],
        "operationId": "find_by_context",
        "parameters": [
          {
            "name": "context",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Context"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find templates by context",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateV2"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/template/find-by-ids": {
      "get": {
        "tags": [
          "template::routes"
        ],
        "operationId": "find_by_ids",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find By ids",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateV2"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/template/find-by-type": {
      "get": {
        "tags": [
          "template::routes"
        ],
        "operationId": "find_by_type",
        "parameters": [
          {
            "name": "template_type",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TemplateType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find templates by type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateV2"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/template/find-one/{templ_id}": {
      "get": {
        "tags": [
          "template::routes"
        ],
        "operationId": "find_one",
        "parameters": [
          {
            "name": "templ_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find a template by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateV2"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/template/render": {
      "post": {
        "tags": [
          "template::routes"
        ],
        "operationId": "render",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Render template",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "contentMediaType": "*/*"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/template/{id}": {
      "delete": {
        "tags": [
          "template::routes"
        ],
        "operationId": "delete_templ_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a template by id"
          }
        }
      }
    },
...
    "/api/v1/upload/{id}": {
      "delete": {
        "tags": [
          "upload::routes"
        ],
        "operationId": "delete_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a file by id"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Context": {
        "type": "string",
        "enum": [
          "INVOICE",
          "PEPPOL",
          "C_V"
        ]
      },
...
      "RenderRequest": {
        "type": "object",
        "required": [
          "templateId",
          "context",
          "fileName",
          "templateContext"
        ],
        "properties": {
          "context": {},
          "fileName": {
            "type": "string"
          },
          "templateContext": {
            "$ref": "#/components/schemas/Context"
          },
          "templateId": {
            "type": "string"
          }
        }
      },
      "TemplateType": {
        "type": "string",
        "enum": [
          "HTML",
          "XML"
        ]
      },
...
      "TemplateV2": {
        "type": "object",
        "required": [
          "_id",
          "creationDate",
          "fileId",
          "templateType",
          "templateContext",
          "title"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "fileId": {
            "type": "string"
          },
          "templateContext": {
            "$ref": "#/components/schemas/Context"
          },
          "templateType": {
            "$ref": "#/components/schemas/TemplateType"
          },
          "title": {
            "type": "string"
          },
          "updatedDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
    }
  }
}